[kcms/effectsmodel] Add config module role to effectsmodel and use it to open config
For builtin effects the information comes from the EffectData struct.
For plugin effects KPluginMetaData::configModule is used if available.
This is added in https://invent.kde.org/frameworks/kcoreaddons/-/merge_requests/32.
For plugins that do not yet make use of it we fall back to the old way.
For scripted effects this is empty since the config is loaded in a different way.